[refactor] App: move storydetail domain models
