+++ app/src/main/java/uk/ryanwong/skycatnews/storydetail/ui/StoryDetailUIState.kt
@@ -5,7 +5,7 @@
 package uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.storydetail.ui
 
 import uk.ryanwong.skycatnews.app.util.ErrorMessage
-import uk.ryanwong.skycatnews.storydetail.domain.model.Story
+import uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.domain.model.storydetail.Story
 
 data class StoryDetailUIState(
     val story: Story? = null,
