+++ app/src/test/java/uk/ryanwong/skycatnews/storydetail/domain/model/StoryTest.kt
@@ -7,6 +7,8 @@ package uk.ryanwong.skycatnews.storydetail.domain.model
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
 import uk.ryanwong.skycatnews.storydetail.data.local.entity.ContentEntity
+import uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.domain.model.storydetail.Content
+import uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.domain.model.storydetail.Story
 
 internal class StoryTest : FreeSpec() {
